final fixxes and cleanups

+++ application/src/main/java/restcontroller/ExerciseController.java
@@ -1,37 +1,32 @@
 package restcontroller;
 
-import entitys.Device;
 import entitys.Exercise;
-import entitys.Muscle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import repositories.DeviceRepository;
 import repositories.MuscleRepository;
 import restcontroller.util.ExerciseDTO;
+import restcontroller.util.ExerciseMapper;
 import restcontroller.util.ExerciseRequestDTO;
 import restcontroller.util.MusclesAndDevicesRequest;
-import services.DeviceService;
 import services.ExerciseService;
-import services.MuscleService;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/exercises")
 @CrossOrigin(origins = "http://localhost:4200", methods = {RequestMethod.GET, RequestMethod.POST, RequestMethod.OPTIONS})
 public class ExerciseController {
 
-    @Autowired
-    private ExerciseService exerciseService;
+    private final ExerciseService exerciseService;
 
-    public ExerciseController(MuscleRepository muscleRepository) {
+    private final ExerciseMapper exerciseMapper;
+    public ExerciseController(ExerciseService exerciseService, ExerciseMapper exerciseMapper) {
+        this.exerciseMapper = exerciseMapper;
+        this.exerciseService = exerciseService;
     }
 
     @GetMapping(produces = "application/json")
@@ -43,21 +38,26 @@ public class ExerciseController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Exercise> getExerciseById(@PathVariable Long id) {
-        Optional<Exercise> exercise = exerciseService.getExerciseById(id);
+        Optional<Exercise> exercise = Optional.ofNullable(exerciseService.getExerciseById(id));
         return exercise.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping
     public ResponseEntity<Exercise> addExercise(@RequestBody ExerciseRequestDTO exerciseRequestDTO, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
-            return ResponseEntity.badRequest().body(null);
+            return ResponseEntity.badRequest().body((Exercise) bindingResult.getAllErrors());
         }
-        Exercise newExercise = ExerciseDTO.getExercise(exerciseRequestDTO);
-        return new ResponseEntity<>(newExercise, HttpStatus.CREATED);
+        Exercise exercise = exerciseMapper.toEntity(exerciseRequestDTO);
+        Exercise savedExercise = exerciseService.addExercise(exercise);
+        return new ResponseEntity<>(savedExercise, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteExercise(@PathVariable Long id) {
+        Optional<Exercise> exercise = Optional.ofNullable(exerciseService.getExerciseById(id));
+        if (exercise.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
         exerciseService.deleteExercise(id);
         return ResponseEntity.noContent().build();
     }
